Some code cleanup - Use Path.Combine instead of inefficiently adding strings - Rename "versionNumber" to "VersionString" and make it constant - Improved readability by adding whitespace - Re-arranged some stuff to Program.cs; allowed the deletion of the FormClosing event amongst other things - Re-worked the temporary file stuff from the manual approach using a sessionID to Microsoft's official TempFileCollection class. -- (This also fixes a bug where temporary stuff wasn't getting deleted)
There seems to be a bug with the application (most likely in the ParseSP2 method) where certain file streams (usually the stream to the first file often called 0.dds) don't get closed. This bug was already present previously, but now with the re-working of the temporary file handling it's much more apparent. All you have to do to see the bug is open a .gpu.sp2 file, then open another .gpu.sp2 file without closing the application. The application throws an exception as it's trying to overwrite the first file (again, usually 0.dds) but can't as the stream is still open.

I plan on fixing this in a future commit where I plan to rework the ParseSP2 method.